+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Update packages
+
 ## 2.8.8 (2023-05-05)
 
 * (theimo1221) Fix prettier error